pushing

+++ ft_floating_types.c
@@ -6,7 +6,7 @@
 /*   By: adaly <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/05/30 14:21:01 by adaly             #+#    #+#             */
-/*   Updated: 2017/06/04 03:34:25 by adaly            ###   ########.fr       */
+/*   Updated: 2017/06/04 03:35:02 by adaly            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -145,7 +145,7 @@ int	ft_floating_types(t_pfconv *current, va_list args)
 	if (ft_lowercase(current->type) == 'a')
 		base = 16;
 	if (current->length == 4)
-		new = va_arg(args, double);
+		new = va_arg(args, long double);
 	else
 		new = va_arg(args, double);
 	if (ft_lowercase(current->type) == 'g')
